+++ .github/workflows/publish.yml
@@ -26,7 +26,7 @@ jobs:
           echo "PR=$(cat env | grep "prev.event.number" | awk -F '=' '{print $2}')" >> $GITHUB_ENV
           echo "BUILD_URL=$(echo ${{ github.repository }} | sed -r 's/\//\.github\.io\//g')" >> $GITHUB_ENV
 
-      - name: Deploy build to Github Pages
+      - name: Deploy build to Github pages
         uses: JamesIves/github-pages-deploy-action@v4
         with:
           folder: dist
