+++ markup/Readme.md
@@ -1,10 +1,10 @@
 # Директория с вёрсткой (markup)
 
 * `index.html` - контейнер;
-* `login.html` - страница «Login» (Авторизация);
-* `main.html` - страница «Main» (Главная страница со списком предложений):
+* `login.html` - страница «LoginPage» (Авторизация);
+* `main.html` - страница «MainPage» (Главная страница со списком предложений):
    * `main-empty.html` - без списка предложений;
-* `offer.html` — страница «Offer» (Карточка предложения по аренде, пользователь авторизован):
+* `offer.html` — страница «OfferPage» (Карточка предложения по аренде, пользователь авторизован):
    * `offer-not-logged.html` - пользователь не авторизован;
-* `favorites.html` - страница «Favorites» (Избранное):
+* `favorites.html` - страница «FavoritesPage» (Избранное):
    * `favorites-empty.html` - пустая страница.
