--- src/Component/Review/Review.tsx
@@ -1,39 +1,0 @@
-import {UserReview} from '../../Mocks/Mocks.ts';
-
-type ReviewProps = {
-  review: UserReview;
-}
-
-function Review({review}: ReviewProps) {
-  return (
-    <li className="reviews__item">
-      <div className="reviews__user user">
-        <div className="reviews__avatar-wrapper user__avatar-wrapper">
-          <img className="reviews__avatar user__avatar" src={review.avatar} width="54" height="54"
-            alt="Reviews avatar"
-          />
-        </div>
-        <span className="reviews__user-name">{review.name}</span>
-      </div>
-      <div className="reviews__info">
-        <div className="reviews__rating rating">
-          <div className="reviews__stars rating__stars">
-            <span style={{width: '80%'}}></span>
-            <span className="visually-hidden">Rating</span>
-          </div>
-        </div>
-        <p className="reviews__text">
-          {review.text}
-        </p>
-        <time className="reviews__time" dateTime={review.date.toISOString()}>
-          {review.date.toLocaleDateString('en-US', {
-            year: 'numeric',
-            month: 'long'
-          })}
-        </time>
-      </div>
-    </li>
-  );
-}
-
-export default Review;
