--- src/Hooks/UseMap.ts
@@ -1,38 +1,0 @@
-import {useEffect, useState, MutableRefObject, useRef} from 'react';
-import {Map, TileLayer} from 'leaflet';
-import Offer from '../Types/Offer.ts';
-
-function useMap(
-  mapRef: MutableRefObject<HTMLElement | null>,
-  {latitude, longitude}: Offer): Map | null {
-
-  const [map, setMap] = useState<Map | null>(null);
-  const isRenderedRef = useRef<boolean>(false);
-
-  useEffect(() => {
-    if (mapRef.current !== null && !isRenderedRef.current) {
-      const instance = new Map(mapRef.current, {
-        center: {
-          lat: latitude,
-          lng: longitude
-        },
-        zoom: 11,
-      });
-
-      instance.addLayer(new TileLayer(
-        'https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png',
-        {
-          attribution:
-            '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>',
-        }
-      ));
-
-      setMap(instance);
-      isRenderedRef.current = true;
-    }
-  }, [mapRef, latitude, longitude]);
-
-  return map;
-}
-
-export default useMap;
