--- src/Pages/FavoritesPage/FavoritesPage.tsx
@@ -1,24 +1,0 @@
-import FavoritesList from '../../Component/FavoritesList/FavoritesList.tsx';
-import Offer from '../../Types/Offer.ts';
-
-function FavoritesPage({offers}: { offers: Offer[] }) {
-  return (
-    <>
-      <main className="page__main page__main--favorites">
-        <div className="page__favorites-container container">
-          <section className="favorites">
-            <h1 className="favorites__title">Saved listing</h1>
-            <FavoritesList offers={offers}/>
-          </section>
-        </div>
-      </main>
-      <footer className="footer container">
-        <a className="footer__logo-link" href="main.html">
-          <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33"/>
-        </a>
-      </footer>
-    </>
-  );
-}
-
-export default FavoritesPage;
