+++ src/components/favorites-card/favorites-card.tsx
