+++ src/components/map/map-component.tsx
@@ -1,0 +1,43 @@
+import { useRef, useEffect } from 'react';
+import { Icon, Marker, layerGroup } from 'leaflet';
+import 'leaflet/dist/leaflet.css';
+import {Offer} from '../../mocks/mocks.ts';
+import useMap from '../../hooks/use-map.ts';
+
+type MapProps = {
+  offers: Offer[];
+  selectedOffer: Offer;
+};
+
+const defaultCustomIcon = new Icon({
+  iconUrl: 'https://assets.htmlacademy.ru/content/intensive/javascript-1/demo/interactive-map/pin.svg',
+  iconSize: [40, 40],
+  iconAnchor: [20, 40],
+});
+
+function Map({offers, selectedOffer}: MapProps){
+  const mapRef = useRef(null);
+  const map = useMap(mapRef, selectedOffer);
+
+  useEffect(() => {
+    if (map) {
+      const markerLayer = layerGroup().addTo(map);
+      offers.forEach((offer) => {
+        const marker = new Marker({
+          lat: offer.latitude,
+          lng: offer.longitude,
+        });
+
+        marker.setIcon(defaultCustomIcon).addTo(markerLayer);
+      });
+
+      return () => {
+        map.removeLayer(markerLayer);
+      };
+    }
+  }, [map, offers, selectedOffer]);
+
+  return <div style={{ height: '500px' }} ref={mapRef}></div>;
+}
+
+export default Map;
