+++ src/constants/links.ts
@@ -1,0 +1,3 @@
+export const MAP_LINK = 'https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png';
+export const OPEN_STREET_MAP_COPYRIGHT = 'https://www.openstreetmap.org/copyright';
+export const CARTO_COPYRIGHT = 'https://carto.com/attributions';
