+++ src/pages/favorites-empty/favorites-empty-page.tsx
