+++ src/types/offer.ts
